Added serverExtended file.
